requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
google-cloud-vision==3.5.0
Pillow==10.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pytest==8.0.0
tailwindcss==3.4.1